Allow up to 6 levels of headers in the table of content and optional ignoring of h1 headers
Any number of levels is supported but since the official HTML
supports only 6 levels (`h1` ... `h6`) an artifical limitation was added.

This change affects two options: `toc_level` and `ordered_chapter_level`.
Both can be set to a value between 0 and 6 (inclusive) where
0 means the feature is disabled.

If either of the options is set to a value greater than 6 then
6 levels of headers are processed and a warning is emitted.

This change also adds option `ignore_top_header` to exclude h1
headers from the table of content and numbering.